+++ app/Http/Controllers/UserController.php
@@ -15,30 +15,28 @@ class UserController extends Controller
     
         }
     }
-    public function  themMoi()
+    public function themMoi()
     {
-        {
-            
-            $dsUser = User::all();
-            return view('user.add',compact('dsUser'));
-        }
+        $dsUser = User::all();
+        return view('user.add', compact('dsUser'));
     }
+
     public function xuLyThemMoi(Request $request)
     {
-        {
-            $user = new User();
-            $user->account_name = $request->account_name;
-            $user->password = $request->password;
-            $user->fullname = $request->fullname;
-            $user->email = $request->email;
-            $user->phone = $request->phone;
-            $user->address = $request->address;
-            $user->address = $request->address;
-            $user->save();
-           
-            return redirect()->route('index')->with(['themMoi'=>"Thêm mới thành công"]);
-        }
+        $user = new User();
+        $user->account_name = $request->account_name;
+        $user->password = $request->password;
+        $user->fullname = $request->fullname;
+        $user->email = $request->email;
+        $user->phone = $request->phone;
+        $user->address = $request->address;
+        $user->save();
+
+        return redirect()->route('user.index')->with(['themMoi' => "Thêm mới thành công"]);
     }
+
+   
+    
     public function  capNhat($id)
     {
         {
@@ -52,7 +50,7 @@ class UserController extends Controller
         }
     }
     public function xuLyCapNhat(Request $request,$id){
-        $admin = User::find($id);
+        $user = User::find($id);
         if(empty($user)){
             return redirect()->back()->withErrors(['loiCapNhap'=>"Sản phẩm không tồn tại"]);
         }
@@ -66,12 +64,23 @@ class UserController extends Controller
         return redirect()-> action([UserController::class, 'index'],['id'=>$user->id])->with(['capNhap'=>"Cập nhật Admin thanh cong"]);
     }
 
-    public function xoa(Request $request,$id){
-
-        $user= User::find($id); 
-        User::where('id', $user->id)->delete();  //Xóa chi tiết sản phẩm liên quan         
-        $user -> delete();
-        return redirect()->route('index')->with(['xoa'=>"xoa Admin thanh cong"]);
-    } 
+    public function xoa(Request $request, $id)
+    {
+        $user = User::find($id);
+    
+        if ($user !== null) {
+            // Xóa chi tiết sản phẩm liên quan
+            User::where('id', $user->id)->delete();
+    
+            // Xóa user
+            $user->delete();
+    
+            return redirect()->route('user.index')->with(['xoa' => "Xóa user thành công"]);
+        } else {
+            // Xử lý khi không tìm thấy user
+            return redirect()->route('user.index')->with(['error' => "User không tồn tại"]);
+        }
+    }
+    
    
 }
